Add reasoning details to ScoreQuiz
